// src/@types/schema/v1/models/GetAccount200Response.d.ts
/**
 * Symbol REST Wrapper
 * API server that makes development more convenient
 *
 * The version of the OpenAPI document: 0.0.1
 * Contact: ym.u.ichiro@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import type { Account } from './Account.js';
/**
 *
 * @export
 * @interface GetAccount200Response
 */
export interface GetAccount200Response {
  /**
   *
   * @type {Account}
   * @memberof GetAccount200Response
   */
  account: Account;
}
/**
 * Check if a given object implements the GetAccount200Response interface.
 */
export declare function instanceOfGetAccount200Response(value: object): boolean;
export declare function GetAccount200ResponseFromJSON(json: any): GetAccount200Response;
export declare function GetAccount200ResponseFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean
): GetAccount200Response;
export declare function GetAccount200ResponseToJSON(value?: GetAccount200Response | null): any;